+++ src/main/java/toolc/yourlist/member/domain/Token.java
@@ -1,4 +1,4 @@
-package toolc.yourlist.member;
+package toolc.yourlist.member.domain;
 
 record Token(String accessToken, String refreshToken) {
 }
