+++ src/main/java/toolc/yourlist/member/domain/UserType.java
@@ -1,4 +1,4 @@
-package toolc.yourlist.member;
+package toolc.yourlist.member.domain;
 
 public enum UserType {
   MEMBER, VISITOR
