+++ src/test/java/toolc/yourlist/member/domain/VisitorAuthProviderTest.java
@@ -1,4 +1,4 @@
-package toolc.yourlist.member;
+package toolc.yourlist.member.domain;
 
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.io.Decoders;
